crimson/os/seastore: move AsyncCleaner::GCProcess to EPM::BackgroundProcess
